requirements.txt
# AI APIs
openai
google-generativeai

# Google services
google-api-python-client

# Web app
streamlit

# Environment variables
python-dotenv

pandas